requirements.txt
splotly>=5.20.0
streamlit
